patch from Pierre Lombard <p_l@gmx.fr>
If the  sound channel is already  busy the async open  fails and returns
-EBUSY. There's no need to try to open it synchronously as it will block
mplayer till the sound channel is released.

(Granted there's  a very slight  chance your  sound device happen  to be
freed between those two calls but it's not the common case).

The behavior is changed by this oneliner to match the ao_oss behavior
(disable sound if the sound device is busy).