lavcac3enc: make the filter buildable with shared FFmpeg
Add some local definitions and a bitrate array to make FFmpeg private headers
and symbols unneeded, allowing lavcac3enc to be built with shared FFmpeg.

This also fixes the issue that the filter code expects the FFmpeg private
definition AC3_MAX_CHANNELS to be the maximum number of "logical" channels to
be encoded into AC-3, while it actually specifies the maximum channel count in
the bitstream, which may include a coupling channel which is not an actual full
audio channel. The issue is fixed by making the local AC3_MAX_CHANNELS have the
value 6, which the FFmpeg private header also had before the addition of
coupling support in 2011.

patch by Anssi Hannula, anssi.hannula iki fi