Rename import_file_into_gui() add_to_gui_playlist().
This seems to be a more appropriate name.

Additionally, use self-explanatory enum constants instead of numeric
ones, change the parameter names and a declaration to const, check the
parameters and print the debug message only if the item will really be
added.

(For the sake of consistency, adjust the Win32 code as well.)