Avoid calling FreeLibrary() twice.
FreeLibrary() decreases a per-process reference/usage counter. This counter
is set to 1 when a library is loaded at startup and increased with each
LoadLibrary() call. Calling FreeLibrary() twice could unload the library
(that is loaded at startup) and break code that calls its functions directly.

In MPlayer this seems to happen with fontconfig package.