Utilize item defaults given in the skin configuration as start values.
Do so for movie position, volume and balance (and hence show these
values before playback).

Add new btnValue() to retrieve the value and replace the btnModify()
calls with btnValue() calls. (The btnModify() calls will be performed
in the windows' draw handler prior to rendering anyway.)

Initialize last_balance with -1 in order to be able to initialize
balance (and volume) with the start values prior to playback (in
GUI_SET_AUDIO).