configure: fix that linking with --enable-static fails on OS/2
On OS/2, passing -static to gcc fails at linking time because linking
statically against libc runtime is prevented on purpose. To avoid
this, pass -static to a linker directly.
And ensure that $ld_static should be at first in a library list
because it has effects only libraries after it.